requirements.txt
paho.mqtt
art
pandas
sqlite3

